Moved scraping into a child process
